+++ NEWS.md
@@ -7,8 +7,8 @@ News
 1.  New functions: getConceptSetsFromCohortDefinition
 
 ### Bug fixes
-1.  https://github.com/OHDSI/ROhdsiWebApi/pull/144 was causing duplication of date matches.
-
+1. https://github.com/OHDSI/ROhdsiWebApi/pull/144 was causing duplication of date matches.
+2. RJSONIO::toJSON loss of precision #152
 
 # ROhdsiWebApi 1.0.0